tiny_Startgate
Version 1.0
Due to the RAM limits of the ATTiny85 only strips up to 90 pixels long can be used.

No further developments on this project are likely at this time. A similar project will be created using an ATMega328 instead of the ATTiny85 to take advantage of the large RAM size.